% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlview.R
\name{highlight_styles}
\alias{highlight_styles}
\title{List available styles}
\usage{
highlight_styles()
}
\description{
Returns a character vector of available style sheets to use when displaying
an XML document.
}
\references{
See \url{https://highlightjs.org/static/demo/} for a demo of all
            highlight.js styles
}

