% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderXmlview}
\alias{renderXmlview}
\title{Widget render function for use in Shiny}
\usage{
renderXmlview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
Widget render function for use in Shiny
}

