% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlview.R
\name{xml_view}
\alias{xml_view}
\title{XML pretty printer & viewer}
\usage{
xml_view(doc, style = "default", width = NULL, height = NULL)
}
\arguments{
\item{doc}{\code{xml2} document, node, nodeset or atomic character vector
of HTML/XML content}

\item{style}{CSS stylesheet to use (see \code{higlight_styles()})}

\item{width}{width}

\item{height}{height}
}
\description{
This uses \code{vkbeautify} and \code{highlight.js} to format and
"pretty print" HTML/XML documents, nodes, node sets and plain character
HTML/XML in an \code{htmlwidget} pane.
}
\examples{
xml_view("<html><head><title>Hi</title></head><body>Test</body></html>")
}
\references{
\href{https://highlightjs.org/}{highlight.js},
            \href{http://www.eslinstructor.net/vkbeautify/}{vkbeautify}
}

