% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlview-package.R
\docType{package}
\name{xmlview}
\alias{xmlview}
\alias{xmlview-package}
\title{Format and Pretty Print 'HTML'/'XML' Content}
\description{
This uses \code{vkbeautify} and \code{highlight.js} to format and
"pretty print" HTML/XML documents, nodes, node sets and plain character
HTML/XML in an \code{htmlwidget} pane.
}
\author{
Bob Rudis (@hrbrmstr)
}

