% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlview.R
\name{xml_view}
\alias{xml_view}
\title{XML pretty printer & viewer}
\usage{
xml_view(doc, style = "default", add_filter = FALSE)
}
\arguments{
\item{doc}{\code{xml2} document, node, nodeset or atomic character vector
of HTML/XML content}

\item{style}{CSS stylesheet to use (see \code{higlight_styles()})}

\item{add_filter}{show an XPath input box to enable live filtering?
(default: \code{FALSE})}
}
\description{
This uses \code{vkbeautify} and \code{highlight.js} to format and
"pretty print" HTML/XML documents, nodes, node sets and plain character
HTML/XML in an \code{htmlwidget} pane.
}
\examples{
library(xml2)
txt <- paste0("<note><to>Tove</to><from>Jani</from><heading>Reminder</heading>",
              "<body>Don't forget me this weekend!</body></note>"
xml_view(txt)
doc <- read_xml(txt)
xml_view(doc, style="obsidian")
xml_view(xml_find_all(doc, ".//to"), style="github-gist")
xml_view(read_xml(system.file("extdata/dwml.xml", package="xmlview")))
xml_view(read_xml(system.file("extdata/getHistory.xml", package="xmlview")),
         "androidstudio")
xml_view(read_xml(system.file("extdata/input.xml", package="xmlview")),
         "sunburst")
}
\references{
\href{https://highlightjs.org/}{highlight.js},
            \href{http://www.eslinstructor.net/vkbeautify/}{vkbeautify}
}

